# Added helper function call_with_deduplication and use it to speed up path_file and path_dir for vectors with repeats.