@gaborcsardi 2025-05-25
It it fine to keep it in my opinion.